split models/lwdetr.py into criterion, postprocess, and math modules
- Extract SetCriterion + loss functions to models/criterion.py
- Extract PostProcess to models/postprocess.py
- Extract MLP + re-export inverse_sigmoid/accuracy/interpolate to models/math.py
- Create models/heads/__init__.py stub (detection head extraction deferred)
- Keep backward-compatible re-exports in lwdetr.py so all existing import paths work
- Update models/__init__.py to re-export from canonical submodule locations
