+++ src/rfdetr/models/__init__.py
@@ -13,4 +13,7 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
 # ------------------------------------------------------------------------
 
-from rfdetr.models.lwdetr import PostProcess, build_criterion_and_postprocessors, build_model
+from rfdetr.models.criterion import SetCriterion  # noqa: F401
+from rfdetr.models.lwdetr import build_criterion_and_postprocessors, build_model
+from rfdetr.models.math import MLP  # noqa: F401
+from rfdetr.models.postprocess import PostProcess  # noqa: F401
