+++ src/rfdetr/models/heads/__init__.py
@@ -1,0 +1,7 @@
+# ------------------------------------------------------------------------
+# RF-DETR
+# Copyright (c) 2025 Roboflow. All Rights Reserved.
+# Licensed under the Apache License, Version 2.0 [see LICENSE for details]
+# ------------------------------------------------------------------------
+
+"""Detection and segmentation head subpackage (stub -- contents extracted in a future phase)."""
